x86-64: make COM1 mandatory (not a lying option); fix UB pointer compare
Two more Copilot review findings on PR #350:

- bios/Kconfig: CONF_WITH_PC_COM1's "n" setting didn't actually disable
  anything. bios/machine/pc-x86_64/earlycon.c (the pre-bios_init()
  diagnostic console startup.c and the panic paths call unconditionally)
  has no #if CONF_WITH_PC_COM1 guard at all, and earlycon.o is linked
  whenever MACHINE_PC_X86_64 is set (bios/build.mk), independent of this
  symbol. Selecting "n" would leave boot completely silent -- this
  machine has no framebuffer yet (#332) -- and, on hardware/emulation
  where an unmapped I/O port doesn't float high, hang outright the
  moment earlycon_putc() busy-waits forever on a THRE bit that never
  sets. Rather than build out a whole disable/no-op early-console path
  for an option nothing needs yet, dropped the user-facing prompt (the
  symbol keeps its default y, help text, and its effect wherever it's
  tested elsewhere -- just no way to flip it in menuconfig): COM1 is
  genuinely mandatory equipment for this machine right now, not a real
  configuration choice, so the option shouldn't offer to do something it
  can't.

- vdi/vdi_raster.c: bit_blt()'s `if ((s_addr < d_addr) || ...)` ordering
  check is undefined behavior now that s_addr/d_addr are UBYTE* (fixed
  in the previous round's truncation fix) -- C only defines `<`/`>`
  between pointers into the same array/object, and these two generally
  point into distinct memory forms (source and destination). Compared
  as (uintptr_t) instead for this one ordering test; the `==` case right
  after is untouched, since pointer equality is always well-defined
  regardless of which objects the operands point into.

Verified: rebuilt and boot-tested pc-x86_64_defconfig under QEMU/OVMF
(CONF_WITH_PC_COM1 confirmed still 1 in obj/autoconf.h despite having no
prompt now; zero warnings from vdi_raster.c; one boot attempt, same
gouser() milestone as before). Compile-tested virt-arm_defconfig (zero
warnings; this is the config that actually exercises bit_blt() live).

+++ vdi/vdi_raster.c
@@ -633,7 +633,17 @@ static void bit_blt(struct blit_frame *blit_info)
            + (ULONG)d_xmin_off * (ULONG)blit_info->d_nxwd);
 
     /* if (just_screen && (s_addr < d_addr)) { */
-    if ((s_addr < d_addr)
+    /* Compared as integer addresses, not as pointers: s_addr/d_addr
+     * generally point into two distinct memory forms (source and
+     * destination), and C only defines `<`/`>` between pointers into the
+     * same array/object -- comparing across two unrelated allocations is
+     * undefined behavior, unlike the equality test just below (pointer
+     * equality is always well-defined, regardless of which objects the
+     * operands point into). This is purely an ordering/overlap check
+     * (mirroring the original ULONG-address comparison from before these
+     * were pointers), not a real array-bounds computation, so reducing
+     * to uintptr_t here costs nothing. */
+    if (((uintptr_t)s_addr < (uintptr_t)d_addr)
      || ((s_addr == d_addr) && (skew >= 0))) {
         /* start from lower right corner, so add width+length */
         s_addr = (UBYTE *)blit_info->s_form
